requirements.txt
aiohttp==3.7.4.post0
alabaster==0.7.12
appdirs==1.4.4
astroid==2.6.6
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
black==21.7b0
bleach==4.0.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
coverage==5.5
cryptography==3.4.7
docutils==0.17.1
idna==3.2
imagesize==1.2.0
importlib-metadata==4.6.3
iniconfig==1.1.1
isort==5.9.3
jeepney==0.7.1
Jinja2==3.0.1
keyring==23.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
multidict==5.1.0
mypy-extensions==0.4.3
packaging==21.0
pathspec==0.9.0
pep8==1.7.1
pkginfo==1.7.1
pluggy==0.13.1
py==1.10.0
pycparser==2.20
Pygments==2.9.0
pylint==2.9.6
pylint-mccabe==0.1.3
pyparsing==2.4.7
pytest==6.2.4
pytest-black==0.3.12
pytest-cov==2.12.1
pytest-runner==5.3.1
pytz==2021.1
readme-renderer==29.0
regex==2021.8.3
requests==2.26.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
SecretStorage==3.3.1
six==1.16.0
# Editable Git install with no remote (slack-bot==0.1.0)
-e /home/vladimir/projects/slack_bot
slackclient==2.9.3
snowballstemmer==2.1.0
Sphinx==4.1.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
toml==0.10.2
tomli==1.2.1
tqdm==4.62.0
twine==3.4.2
typing-extensions==3.10.0.0
urllib3==1.26.6
webencodings==0.5.1
wrapt==1.12.1
yarl==1.6.3
zipp==3.5.0
